+++ src/schematics/schematics.jl
@@ -1325,6 +1325,9 @@ function build!(sch::Schematic, geometry_fn=geometry; strict=:error)
             node_cs = structure(ref)
             # One of the node cs's references will be its Component
             for (i, subref) in pairs(refs(node_cs)) # Just check all of them
+                if structure(subref) isa AbstractCompositeComponent
+                    build!(schematic(structure(subref)))
+                end
                 if structure(subref) isa AbstractComponent
                     try
                         refs(node_cs)[i] =